fix(ui): merge shimmer into flex widget
