+++ lib/product_list/product_list_page.dart
@@ -2,14 +2,12 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flex_storefront/flex_ui/components/app_bar.dart';
 import 'package:flex_storefront/flex_ui/tokens/sizes.dart';
 import 'package:flex_storefront/flex_ui/widgets/product_list_item.dart';
-import 'package:flex_storefront/flex_ui/widgets/product_list_item_shimmer.dart';
 import 'package:flex_storefront/flex_ui/widgets/search_results_header.dart';
 import 'package:flex_storefront/product_list/cubits/product_search_cubit.dart';
 import 'package:flex_storefront/product_list/cubits/product_search_state.dart';
 import 'package:flex_storefront/shared/bloc_helper.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter/material.dart';
-import 'package:shimmer/shimmer.dart';
 
 @RoutePage()
 class ProductListPage extends StatelessWidget {
